requirements.txt
torch>=2.0.0
psycopg2-binary>=2.9.9
pandas>=2.0.0
numpy>=1.24.0
transformers>=4.30.0
datasets>=2.12.0
scikit-learn>=1.2.0
tqdm>=4.65.0
python-dotenv>=1.0.0 